DC-457 GL updates

+++ DDI/DDI.Shared/Models/Client/GL/FiscalYear.cs
@@ -28,20 +28,14 @@ namespace DDI.Shared.Models.Client.GL
         public FiscalYearStatus Status { get; set; }
 
         public int NumberOfPeriods { get; set; }
-
-        // JLT Why a byte?
-        public byte CurrentPeriodNumber { get; set; }
+         
+        public int CurrentPeriodNumber { get; set; }
 
         public ICollection<FiscalPeriod> FiscalPeriods { get; set; }
 
         public ICollection<LedgerAccount> LedgerAccounts { get; set; }
 
         public ICollection<Segment> Segments { get; set; }
-        
-        // JFA:
-        // CurrentPeriod is definitely required.
-        // NumberOfPeriods may be helpful to avoid having to drill into FiscalPeriod to count them.
-        // JT: How does Current Period and Number of Periods get updated or stay up to date if they are 
-        // stored as a field value?
+         
     }
 }
